Update best_answer_id on questions table when answer removed in database layer

+++ database/migrations/2020_01_22_192918_create_answers_table.php
@@ -15,7 +15,7 @@ class CreateAnswersTable extends Migration
     {
         Schema::create('answers', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('question_id');
+            $table->unsignedBigInteger('question_id');
             $table->unsignedBigInteger('user_id');
             $table->text('body');
             $table->integer('votes_count')->default(0);
